requirements.txt
asgiref==3.11.1
Django==6.0.3
sqlparse==0.5.5
mysql-connector-python>=8.0.30
